Refactor: Upgrade to TypeScript NodeNext Module Resolution
- Relative imports now include file extensions
- Merged TypeScript implementation types, removing `*.types.ts` files
- Updated project type to `module` for new resolution strategy
- Reintroduced `phunt-jest-config` for `.js` file support
- Refined package exports using relative paths (e.g., `@112dev/phunt-core/db`)
