+++ apps/phunt-cli/src/commands/index.command.ts
@@ -1,17 +1,17 @@
 import { Command } from "commander";
-import { DefaultMediaFileExtensions } from "../defaults";
+import { DefaultMediaFileExtensions } from "../defaults.js";
 import * as path from "node:path";
 import { promises as fs, Stats } from "node:fs";
 import {
-  WinstonBasedLogger,
-  DateParser,
   FileIndexTableDbService,
   getDefaultDatabaseFileName,
   initDb,
-  LocalFileSystemFileSearchService,
-  FileOps,
-  IndexDestinationDirectoryService,
-} from "@112dev/phunt-core";
+} from "@112dev/phunt-core/db";
+import { WinstonBasedLogger } from "@112dev/phunt-core/logger";
+import { DateParser } from "@112dev/phunt-core/date-parser";
+import { FileOps } from "@112dev/phunt-core/file-ops";
+import { LocalFileSystemFileSearchService } from "@112dev/phunt-core/file-search";
+import { IndexDestinationDirectoryService } from "@112dev/phunt-core/index-dest-dir";
 
 type IndexCommandOptions = {
   ext: string[];
