+++ packages/phunt-contracts/package.json
@@ -2,31 +2,23 @@
   "name": "@112dev/phunt-contracts",
   "version": "1.0.0-beta.2",
   "license": "Apache-2.0",
+  "type": "module",
   "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "default": "./dist/index.js"
-    },
     "./db": {
-      "types": "./src/db.ts",
-      "default": "./dist/db.js"
+      "types": "./dist/db.d.ts"
     },
     "./file-search": {
-      "types": "./src/file-search.ts",
-      "default": "./dist/file-search.js"
+      "types": "./dist/file-search.d.ts"
     },
     "./file": {
-      "types": "./src/file.ts",
-      "default": "./dist/file.js"
+      "types": "./dist/file.d.ts"
     },
     "./logger": {
-      "types": "./src/logger.ts",
-      "default": "./dist/logger.js"
+      "types": "./dist/logger.d.ts"
     }
   },
   "files": [
-    "dist",
-    "src"
+    "dist"
   ],
   "scripts": {
     "build": "tsc -p tsconfig.build.json",
