+++ packages/phunt-contracts/tsconfig.build.json
@@ -1,4 +1,8 @@
 {
   "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "declaration": true,
+    "emitDeclarationOnly": true
+  },
   "exclude": ["src/**/*.spec.ts", "src/**/*.test.ts"]
 }
