--- packages/phunt-core/src/date-parser/index.ts
@@ -1,1 +1,0 @@
-export * from "./date-parser";
