--- packages/phunt-core/src/db/file-index/index.ts
@@ -1,2 +1,0 @@
-export * from "./file-index.types";
-export * from "./file-index";
