+++ packages/phunt-core/src/db/index.ts
@@ -1,2 +1,2 @@
-export * from "./file-index";
-export * from "./init-scripts";
+export * from "./file-index.js";
+export * from "./init-scripts.js";
