--- packages/phunt-core/src/file-ops/index.ts
@@ -1,1 +1,0 @@
-export * from "./file-ops";
