--- packages/phunt-core/src/file-search/index.ts
@@ -1,1 +1,0 @@
-export * from "./file-search";
