--- packages/phunt-core/src/file-sync/file-sync.types.ts
@@ -1,10 +1,0 @@
-export type DuplicateFilterStrategy = "checksum" | "bpb";
-
-export type FileSyncCriteria = {
-  srcFile: string;
-  destDir: string;
-  removeSrc: boolean;
-  destPattern: string;
-  includeDuplicates: boolean;
-  duplicateFilterStrategy: DuplicateFilterStrategy;
-};
