--- packages/phunt-core/src/file-sync/index.ts
@@ -1,2 +1,0 @@
-export * from "./file-sync.types";
-export * from "./file-sync";
