--- packages/phunt-core/src/file-sync/validator/duplicate-file/index.ts
@@ -1,1 +1,0 @@
-export * from "./duplicate-file-validator";
