--- packages/phunt-core/src/file-sync/validator/index.ts
@@ -1,1 +1,0 @@
-export * from "./validator";
