--- packages/phunt-core/src/file-sync/validator/validator.types.ts
@@ -1,5 +1,0 @@
-import { FileSyncCriteria } from "../file-sync.types";
-
-export interface FileSyncValidator {
-  validateAsync(criteria: FileSyncCriteria): Promise<void>;
-}
