--- packages/phunt-core/src/index-dest-dir/index.ts
@@ -1,2 +1,0 @@
-export * from "./index-dest-dir.types";
-export * from "./index-dest-dir";
