--- packages/phunt-core/src/index.ts
@@ -1,7 +1,0 @@
-export * from "./date-parser";
-export * from "./db";
-export * from "./file-ops";
-export * from "./file-search";
-export * from "./file-sync";
-export * from "./index-dest-dir";
-export * from "./logger";
