--- packages/phunt-core/src/logger/index.ts
@@ -1,1 +1,0 @@
-export * from "./winston-logger";
