+++ packages/phunt-core/tsconfig.base.json
@@ -1,7 +1,7 @@
 {
   "extends": "@112dev/phunt-typescript-config/node.tsconfig.json",
   "compilerOptions": {
-    "outDir": "dist"
+    "outDir": "./dist"
   },
   "include": ["src/**/*.ts"]
 }
