+++ packages/phunt-core/tsconfig.build.json
@@ -1,4 +1,7 @@
 {
   "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "declaration": true
+  },
   "exclude": ["src/**/*.spec.ts", "src/**/*.test.ts"]
 }
