+++ packages/phunt-jest-config/base.mjs
@@ -1,0 +1,20 @@
+/**
+ * @type {import("jest").Config}
+ */
+export default {
+  verbose: true,
+  testEnvironment: "node",
+  passWithNoTests: true,
+  transform: {
+    "\\.[jt]s?$": [
+      "ts-jest",
+      {
+        useESM: true,
+      },
+    ],
+  },
+  moduleNameMapper: {
+    "(.+)\\.js": "$1",
+  },
+  extensionsToTreatAsEsm: [".ts"],
+};
