+++ packages/phunt-typeguards/package.json
@@ -2,9 +2,11 @@
   "name": "@112dev/phunt-typeguards",
   "version": "1.0.0-beta.2",
   "license": "Apache-2.0",
+  "type": "module",
+  "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "types": "./src/index.ts",
+      "types": "./dist/index.d.ts",
       "default": "./dist/index.js"
     }
   },
