+++ packages/phunt-typeguards/tsconfig.json
@@ -1,3 +1,3 @@
 {
-  "extends": "tsconfig.base.json"
+  "extends": "./tsconfig.base.json"
 }
